requirements.txt
appdirs==1.4.3
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
dominate==2.5.1
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
win-inet-pton==1.1.0
wincertstore==0.2
WTForms==2.2.1
